# Refactor Environment tests