--- a/tests/mock-constants.php
+++ b/tests/mock-constants.php
@@ -55,3 +55,15 @@ namespace Automattic\VIP\Utils {
 		return Constant_Mocker::constant( $constant );
 	}
 }
+
+namespace Automattic\VIP {
+	use Automattic\Test\Constant_Mocker;
+
+	function defined( $constant ) {
+		return Constant_Mocker::defined( $constant );
+	}
+
+	function constant( $constant ) {
+		return Constant_Mocker::constant( $constant );
+	}
+}