Refactor Environment tests to allow them to run in a single process.

Timings (tests/lib/environment/test-class-environment.php + tests/lib/helpers/test-environment.php):
* before: 51.87 seconds + 17.36 seconds
* after: 2.97 seconds + 2.8 seconds